+++ photorealism-plugin/src/postprocess/postprocess.hpp
@@ -1,10 +1,13 @@
 #pragma once
 
+#include <d3d11.h>
 #include <dxgi.h>
 
 namespace photorealism {
 void process_frame(IDXGISwapChain* swap_chain);
 void upscale_present_frame(IDXGISwapChain* swap_chain);
+void reconstruct_game_frame(
+    ID3D11DeviceContext* context, ID3D11RenderTargetView* output);
 void draw_overlay_frame(IDXGISwapChain* swap_chain);
 bool is_processing_frame();
 void prepare_for_resize(